# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * NyayAI - Indian Legal Assistant API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Conversation {
  id: number;
  title: string;
  /** @nullable */
  topic?: string | null;
  messageCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface ConversationInput {
  title: string;
  topic?: string;
}

export interface Message {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface ConversationWithMessages {
  id: number;
  title: string;
  /** @nullable */
  topic?: string | null;
  createdAt: string;
  updatedAt: string;
  messages: Message[];
}

export interface MessageInput {
  content: string;
}

export interface DocumentField {
  name: string;
  label: string;
  type: string;
  required: boolean;
  placeholder?: string;
}

export interface DocumentType {
  id: string;
  name: string;
  description: string;
  fields: DocumentField[];
}

export interface ExplainInput {
  text: string;
  /** Preferred response language (e.g. en, hi, ta) */
  language?: string;
}

export interface LegalDocument {
  id: number;
  title: string;
  documentType: string;
  content: string;
  /** JSON-encoded field values used to generate this document */
  fields?: string;
  createdAt: string;
}

export type DocumentInputFields = { [key: string]: unknown };

export interface DocumentInput {
  documentType: string;
  title: string;
  fields: DocumentInputFields;
}

export interface LegalStats {
  totalConversations: number;
  totalMessages: number;
  totalDocuments: number;
}

export interface ApiError {
  error: string;
}