# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NyayAI - Indian Legal Assistant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /legal/explain:
    post:
      operationId: explainDocument
      tags: [legal]
      summary: Explain a legal document in plain language
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExplainInput"
      responses:
        "200":
          description: SSE stream of explanation
          content:
            text/event-stream: {}
  /legal/documents:
    get:
      operationId: listDocuments
      tags: [legal]
      summary: List all generated legal documents
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LegalDocument"
    post:
      operationId: createDocument
      tags: [legal]
      summary: Create a legal document from provided details
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentInput"
      responses:
        "200":
          description: SSE stream of generated document
          content:
            text/event-stream: {}
  /legal/documents/{id}:
    get:
      operationId: getDocument
      tags: [legal]
      summary: Get a legal document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Legal document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LegalDocument"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteDocument
      tags: [legal]
      summary: Delete a legal document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    ExplainInput:
      type: object
      properties:
        text:
          type: string
        language:
          type: string
          description: Preferred response language (e.g. en, hi, ta)
      required:
        - text
    LegalDocument:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        documentType:
          type: string
        content:
          type: string
        fields:
          type: string
          description: JSON-encoded field values used to generate this document
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - documentType
        - content
        - createdAt
    DocumentInput:
      type: object
      properties:
        documentType:
          type: string
        title:
          type: string
        fields:
          type: object
          additionalProperties: true
      required:
        - documentType
        - title
        - fields
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error