# lib/api-zod/src/generated/types/documentInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * NyayAI - Indian Legal Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { DocumentInputFields } from './documentInputFields';

export interface DocumentInput {
  documentType: string;
  title: string;
  fields: DocumentInputFields;
}