# lib/api-zod/src/generated/types/explainInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * NyayAI - Indian Legal Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface ExplainInput {
  text: string;
  /** Preferred response language (e.g. en, hi, ta) */
  language?: string;
}