# lib/api-zod/src/generated/types/legalDocument.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * NyayAI - Indian Legal Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface LegalDocument {
  id: number;
  title: string;
  documentType: string;
  content: string;
  /** JSON-encoded field values used to generate this document */
  fields?: string;
  createdAt: Date;
}